GEOS-1417 Styles with many feature type style are working for with attribute filtering disabled. Pumping up a little the number of filters that can be sent to the db.
git-svn-id: http://svn.codehaus.org/geoserver/trunk/geoserver@7654 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: f4256877fe3c3d5e2fc0b4084f281497c45665ee [formerly 3fe4b7ff230c79fe1e8bb1cd1d60e9b8bb1a332b]
Former-commit-id: d15e205a4d21f967335ad8870c397aa65a31343e

+++ wms/src/main/java/org/vfny/geoserver/wms/responses/DefaultRasterMapProducer.java
@@ -302,7 +302,7 @@ public abstract class DefaultRasterMapProducer extends
 		rendererParams.put("optimizedDataLoadingEnabled", new Boolean(true));
 		rendererParams.put("renderingBuffer", new Integer(mapContext
 				.getBuffer()));
-		rendererParams.put("maxFiltersToSendToDatastore", new Integer(10));
+		rendererParams.put("maxFiltersToSendToDatastore", new Integer(20));
 		rendererParams.put(ShapefileRenderer.SCALE_COMPUTATION_METHOD_KEY,
 				ShapefileRenderer.SCALE_OGC);
 		if(AA_NONE.equals(antialias)) {
